Still struggling
So still struggling with the containers for the HLS streams but I'm making some progress. At the moment I'm getting an exception on the ObjectContainer reponse so I haven't even tested if the streams work yet;
AttributeException: If no URL is provided, the key and rating_key attributes must be set.