Everything looks correct
Been through all of the server logs and trans-coding and I can't figure it out. I'll continue to fiddle with it. The parts are type .ts (mpegts) but Plex is not playing them. Just get a loading icon. Only error in transcoder is the following;
TranscodeUniversalRequest: unexpected number of child items: 2

I suspect I need to look at how the MediaObjects are put together again.